branch out separate backend, go full static for frontend
