+++ backend/pyproject.toml
@@ -1,0 +1,28 @@
+[tool.poetry]
+name = "backend"
+version = "0.1.0"
+description = "Scaffold for SIDESTREAM python projects"
+authors = ["SIDESTREAM <team@sidestream.tech>"]
+
+[tool.poetry.dependencies]
+python = "^3.9"
+python-json-logger = "^2.0.1"
+uvicorn = "^0.13.3"
+fastapi = "^0.63.0"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.0"
+pytest-mypy = "^0.8.0"
+pytest-pydocstyle = "^2.2.0"
+flake8-quotes = "^3.2.0"
+flake8-print = "^4.0.0"
+pytest-sugar = "^0.9.4"
+pytest-flake8 = "^1.0.6"
+flake8-import-order = "^0.18.1"
+pytest-xdist = "^2.1.0"
+pytest-cov = "^2.10.1"
+darglint = "^1.5.8"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
