+++ backend/setup.cfg
@@ -1,0 +1,47 @@
+[pydocstyle]
+convention = numpy
+explain = True
+
+[mypy]
+python_version = 3.9
+disallow_untyped_defs = True
+ignore_missing_imports = True
+strict_optional = True
+warn_redundant_casts = True
+warn_return_any = True
+warn_unused_configs = True
+warn_unused_ignores = True
+disallow_incomplete_defs = True
+check_untyped_defs = True
+show_error_context = True
+
+[mypy-tests.*]
+ignore_errors = True
+
+
+[coverage:report]
+fail_under = 95
+show_missing = True
+
+[flake8]
+import-order-style = cryptography
+exclude =
+    __pycache__/
+    .venv/
+    .git/
+max-line-length = 119
+max-complexity = 15
+# [darglint], docstring linting integrated
+# with `flake8`, for strictness levels see
+# [here](https://github.com/terrencepreilly/darglint#strictness-configuration)
+strictness=long
+docstring_style=numpy
+
+[tool:pytest]
+addopts =
+    --cov=src
+    --cov-report term-missing
+    --cov-branch
+    --flake8
+    --pydocstyle
+    --mypy
