+++ backend/src/backend/__init__.py
@@ -1,0 +1,8 @@
+"""Package to provide python back-end scaffolding."""
+from ._logging import configure_logging
+
+
+__version__ = '0.1.0'
+
+
+configure_logging()
