+++ backend/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing tests."""
