+++ backend/tests/test_python_scaffold_version.py
@@ -1,0 +1,6 @@
+"""Module that test the package end to end."""
+from python_scaffold import __version__
+
+
+def test_version():
+    assert __version__ == '0.1.0'
