+++ frontend/Dockerfile
@@ -20,7 +20,11 @@ ENV NODE_ENV production
 ENV NUXT_HOST 0.0.0.0
 ENV NUXT_PORT 3000
 
-# Configure other environment
+# Configure backend
+ARG BACKEND_BASE_URL
+ENV BACKEND_BASE_URL=${BACKEND_BASE_URL}
+
+# Configure other environment settings
 ARG GOOGLE_ID
 ENV GOOGLE_ID=${GOOGLE_ID}
