+++ frontend/README.md
@@ -1,0 +1,18 @@
+# Frontend
+
+Webapp that provides all the UI/UX of `isthisanmlm.com`.
+
+## Development
+
+Install `npm` and `node`, then run:
+```sh
+> npm i
+> npm run dev
+```
+
+to start the application in development mode.
+
+In order to set the backend base url, export:
+```sh
+> export BACKEND_BASE_URL=your_desired_base_url
+```
