+++ frontend/components/FullMLMList.vue
@@ -6,7 +6,7 @@
         <p class="text-center">
             (MLMs are bad and fraudulent compnaies, please stay away from them!)
         </p>
-        <div @click="goto(mlm)" v-for="mlm in mlms" :key="mlm" class="shadow-lg rounded-sm p-4 m-2 bg-white hover:bg-gray-200 cursor-pointer">
+        <div @click="goto(mlm)" v-for="mlm in allMLMs" :key="mlm" class="shadow-lg rounded-sm p-4 m-2 bg-white hover:bg-gray-200 cursor-pointer">
             {{ mlm }}
         </div>
 
@@ -24,9 +24,9 @@ export default {
         }
     },
     computed: {
-        ...mapGetters({
-            mlms: 'getMLMs'
-        })
+        ...mapGetters([
+            'allMLMs'
+        ])
     },
     methods: {
         goto(mlm) {
