+++ frontend/package.json
@@ -38,7 +38,6 @@
     "eslint-plugin-promise": ">=4.0.1",
     "eslint-plugin-standard": "^5.0.0",
     "eslint-plugin-vue": "^5.2.2",
-    "nodemon": "^1.19.4",
     "prettier": "^1.19.1",
     "tailwindcss": "^1.9.6"
   }
