+++ frontend/pages/all-mlms/index.vue
