+++ frontend/plugins/vue-goodshare.js
