+++ frontend/static/favicon.ico
